[dev-cql-parser]Move CQL parser code into its own package
