+++ src/main/java/org/folio/rest/impl/CodexInstancesResourceImpl.java
@@ -5,6 +5,7 @@ import java.util.Map;
 import javax.ws.rs.core.Response;
 
 import org.folio.config.RMAPIConfiguration;
+import org.folio.cql2rmapi.CQLParserForRMAPI;
 import org.folio.rest.annotations.Validate;
 import org.folio.rest.jaxrs.resource.CodexInstancesResource;
